Move test utility classes in a dedicated package
