+++ leshan-integration-tests/src/test/java/org/eclipse/leshan/integration/tests/QueueModeTest.java
@@ -21,6 +21,7 @@ import static org.junit.Assert.*;
 import org.eclipse.leshan.core.Link;
 import org.eclipse.leshan.core.request.ReadRequest;
 import org.eclipse.leshan.core.response.ReadResponse;
+import org.eclipse.leshan.integration.tests.util.QueueModeIntegrationTestHelper;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
