+++ leshan-integration-tests/src/test/java/org/eclipse/leshan/integration/tests/util/SecureIntegrationTestHelper.java
@@ -14,7 +14,7 @@
  *     Zebra Technologies - initial API and implementation
  *******************************************************************************/
 
-package org.eclipse.leshan.integration.tests;
+package org.eclipse.leshan.integration.tests.util;
 
 import java.io.FileInputStream;
 import java.io.IOException;
