+++ leshan-integration-tests/src/test/java/org/eclipse/leshan/integration/tests/util/SynchronousRegistrationListener.java
@@ -13,7 +13,7 @@
  * Contributors:
  *     Sierra Wireless - initial API and implementation
  *******************************************************************************/
-package org.eclipse.leshan.integration.tests;
+package org.eclipse.leshan.integration.tests.util;
 
 import java.util.Collection;
 import java.util.concurrent.CountDownLatch;
